refactor cli transform command
